backend/requirements.txt
asgiref==3.10.0
asttokens==3.0.0
decorator==5.2.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
exceptiongroup==1.3.0
executing==2.2.1
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.14
